Cineflow/requirements.txt
appnope==0.1.4
asgiref==3.8.1
asttokens==3.0.0
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.3
click==8.1.8
cloudinary==1.44.0
comm==0.2.2
crispy-bootstrap5==2025.6
debugpy==1.8.12
decorator==5.1.1
Django==5.2.3
django-cors-headers==4.7.0
django-crispy-forms==2.4
django-formtools==2.5.1
django-otp==1.6.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
executing==2.2.0
Flask==3.1.0
huepy==1.2.1
idna==3.10
ipykernel==6.29.5
ipython==8.32.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jupyter_client==8.6.3
jupyter_core==5.7.2
livereload==2.7.1
markdown2==2.5.4
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
optional-django==0.3.0
packaging==24.2
parso==0.8.4
pexpect==4.9.0
pillow==11.3.0
platformdirs==4.3.6
prompt_toolkit==3.0.50
psutil==7.0.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.1
PyJWT==2.10.1
pypng==0.20220715.0
python-dateutil==2.9.0.post0
pyzmq==26.2.1
qrcode==7.4.2
react==4.3.0
requests==2.32.4
six==1.17.0
sqlparse==0.5.3
stack-data==0.6.3
tornado==6.4.2
traitlets==5.14.3
typing_extensions==4.14.0
urllib3==2.4.0
vite==1.5.2
wcwidth==0.2.13
Werkzeug==3.1.3
